Refactor - move implementation of the discovery context to the package level after the review
