+++ cdi-unit/src/main/java/org/jglue/cdiunit/internal/WeldComponentFactory.java
@@ -1,0 +1,67 @@
+package org.jglue.cdiunit.internal;
+
+import org.jboss.weld.bootstrap.spi.BeanDiscoveryMode;
+import org.jboss.weld.bootstrap.spi.BeansXml;
+import org.jboss.weld.bootstrap.spi.Metadata;
+import org.jboss.weld.bootstrap.spi.Scanning;
+import org.jboss.weld.metadata.BeansXmlImpl;
+
+import java.lang.reflect.Constructor;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.util.ArrayList;
+import java.util.Arrays;
+
+public final class WeldComponentFactory {
+
+	private WeldComponentFactory() {
+	}
+
+	static <T> Metadata<T> createMetadata(T value, String location) {
+		try {
+			return new org.jboss.weld.bootstrap.spi.helpers.MetadataImpl<>(value, location);
+		} catch (NoClassDefFoundError e) {
+			// MetadataImpl moved to a new package in Weld 2.4, old copy removed in 3.0
+			try {
+				// If Weld < 2.4, the new package isn't there, so we try the old package.
+				//noinspection unchecked
+				Class<Metadata<T>> oldClass = (Class<Metadata<T>>) ClassLookup.INSTANCE.lookup("org.jboss.weld.metadata.MetadataImpl");
+				Constructor<Metadata<T>> ctor = oldClass.getConstructor(Object.class, String.class);
+				return ctor.newInstance(value, location);
+			} catch (ReflectiveOperationException e1) {
+				throw new RuntimeException(e1);
+			}
+		}
+	}
+
+	static Object annotatedDiscoveryMode() {
+		try {
+			return BeanDiscoveryMode.ANNOTATED;
+		} catch (NoClassDefFoundError e) {
+			// No such enum in Weld 1.x, but the constructor for BeansXmlImpl has fewer parameters so we don't need it
+			return null;
+		}
+	}
+
+	static BeansXml createBeansXml() {
+		try {
+			// The constructor for BeansXmlImpl has added more parameters in newer Weld versions. The parameter list
+			// is truncated in older version of Weld where the number of parameters is shorter, thus omitting the
+			// newer parameters.
+			Object[] initArgs = new Object[]{
+				new ArrayList<Metadata<String>>(), new ArrayList<Metadata<String>>(),
+				new ArrayList<Metadata<String>>(), new ArrayList<Metadata<String>>(), Scanning.EMPTY_SCANNING,
+				// These were added in Weld 2.0:
+				new URL("file:cdi-unit"), annotatedDiscoveryMode(), "cdi-unit",
+				// isTrimmed: added in Weld 2.4.2 [WELD-2314]:
+				false
+			};
+			Constructor<?> beansXmlConstructor = BeansXmlImpl.class.getConstructors()[0];
+			return (BeansXml) beansXmlConstructor.newInstance(
+				Arrays.copyOfRange(initArgs, 0, beansXmlConstructor.getParameterCount()));
+		} catch (MalformedURLException | ReflectiveOperationException e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+}
